Game Improvement
Now it's also possible to spawn hazards, the hazards are the same fruits that appear in the fruit type
however they are poisoned, so if the player take them, he will get damage. The color of the fruit is changed slightly to green
so it's easier to identify, particles will be also added in the future. The game also has a mini life system, if the players
losels all three lives, whether it's by gathering a poisoned fruit or letting a normal fruit pass, we will lose life.
Further improvements and features will be added!