Game Improvements2
Now it's also possible to gather power ups, there are two in the moment, speed power up and a power up to increase the size of
the fruit gatherer(bowl). The game now also has a CurrencyManager, it's still in working progress, however its goal is to
store the player score and the player money. The upgrade_manager will also be built with the shop upgrades in mind.